+++ RoboUnion/ConfigModel.cs
@@ -11,5 +11,6 @@ internal record class ConfigModel(ConfigFile ConfigFile)
     public readonly ConfigEntry<int> MaxPlayers = ConfigFile.Bind(GeneralSection, "MaxPlayers", 10);
 
     internal const string DeveloperSection = "Developer";
+    public readonly ConfigEntry<bool> LogMessagesPerSecond = ConfigFile.Bind(DeveloperSection, "LogMessagesPerSecond", false);
     public readonly ConfigEntry<bool> ShowMessagesPerSecond = ConfigFile.Bind(DeveloperSection, "ShowMessagesPerSecond", false);
 }