+++ RoboUnion/RoboUnion.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 
 namespace Linkoid.Repo.RoboUnion;
 
-[BepInPlugin("Linkoid.Repo.RoboUnion", "RoboUnion", "0.1")]
+[BepInPlugin("Linkoid.Repo.RoboUnion", "RoboUnion", "0.3")]
 public class RoboUnion : BaseUnityPlugin
 {
     internal static RoboUnion Instance { get; private set; } = null!;