+++ RoboUnion/CHANGELOG.md
@@ -1,1 +1,3 @@
+* 0.3 - Improved and fixed error with photon watchdog
+* 0.2 - Fixed steam lobby size bug 
 * 0.1 - Pre-Release